Support Xcode 7.3. Minimum required OS X version is now 10.9.

+++ Source/CarbonEngine/Game/GameCenter.mm
@@ -398,7 +398,7 @@ void GameCenter::reportAchievementProgress(const String& achievementID, float pe
     }
 
     auto achievement = [[GKAchievement alloc] initWithIdentifier:achievementID.toNSString()];
-    achievement.percentComplete = percentComplete;
+    achievement.percentComplete = double(percentComplete);
 
     auto achievements = pointer_to<NSArray>::type();
